Consolidate parse spec
